# docs/openapi.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/auth/otp/solicitar":{"post":{"operationId":"OtpController_solicitar","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SolicitarOtpDto"}}}},"responses":{"201":{"description":""}},"tags":["OTP"]}},"/auth/otp/validar":{"post":{"operationId":"OtpController_validar","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ValidarOtpDto"}}}},"responses":{"201":{"description":""}},"tags":["OTP"]}},"/auth/otp/reenviar":{"post":{"operationId":"OtpController_reenviar","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReenviarOtpDto"}}}},"responses":{"201":{"description":""}},"tags":["OTP"]}},"/solicitudes":{"post":{"operationId":"SolicitudesController_crear","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrearSolicitudDto"}}}},"responses":{"201":{"description":""}},"security":[{"jwt-ciudadano":[]}],"tags":["Solicitudes"]}},"/solicitudes/mis-solicitudes":{"get":{"operationId":"SolicitudesController_misSolicitudes","parameters":[],"responses":{"200":{"description":""}},"security":[{"jwt-ciudadano":[]}],"tags":["Solicitudes"]}},"/solicitudes/{codigo}":{"get":{"operationId":"SolicitudesController_consultarPorCodigo","parameters":[{"name":"codigo","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Solicitudes"]}},"/admin/solicitudes":{"get":{"operationId":"AdminController_listar","parameters":[],"responses":{"200":{"description":""}},"security":[{"jwt-interno":[]}],"tags":["Admin"]}},"/admin/solicitudes/{id}":{"get":{"operationId":"AdminController_detalle","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt-interno":[]}],"tags":["Admin"]}},"/admin/solicitudes/{id}/asignar":{"patch":{"operationId":"AdminController_asignar","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AsignarOperarioDto"}}}},"responses":{"200":{"description":""}},"security":[{"jwt-interno":[]}],"tags":["Admin"]}},"/admin/solicitudes/{id}/estado":{"patch":{"operationId":"AdminController_cambiarEstado","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CambiarEstadoDto"}}}},"responses":{"200":{"description":""}},"security":[{"jwt-interno":[]}],"tags":["Admin"]}},"/admin/solicitudes/{id}/pdf":{"post":{"operationId":"AdminController_cargarPdf","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"jwt-interno":[]}],"tags":["Admin"]}},"/webhooks/pago":{"post":{"operationId":"WebhookController_recibirPago","parameters":[{"name":"x-pasarela-signature","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Webhooks"]}},"/admin/dashboard":{"get":{"operationId":"DashboardController_stats","parameters":[{"name":"circunscripcion","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"jwt-interno":[]}],"tags":["Dashboard"]}},"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}}},"info":{"title":"RDAM API","description":"Registro de Deudores Alimentarios Morosos - API del Poder Judicial de Santa Fe","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"jwt-interno":{"scheme":"bearer","bearerFormat":"JWT","type":"http"},"jwt-ciudadano":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDto":{"type":"object","properties":{}},"SolicitarOtpDto":{"type":"object","properties":{}},"ValidarOtpDto":{"type":"object","properties":{}},"ReenviarOtpDto":{"type":"object","properties":{}},"CrearSolicitudDto":{"type":"object","properties":{}},"AsignarOperarioDto":{"type":"object","properties":{}},"CambiarEstadoDto":{"type":"object","properties":{}}}}}